Move bayes tree marg helper (no .i presence)
