+++ gtsam/nonlinear/IncrementalFixedLagSmoother.cpp
@@ -20,7 +20,7 @@
  */
 
 #include <gtsam/nonlinear/IncrementalFixedLagSmoother.h>
-#include <gtsam_unstable/nonlinear/BayesTreeMarginalizationHelper.h>
+#include <gtsam/nonlinear/BayesTreeMarginalizationHelper.h>
 #include <gtsam/base/debug.h>
 
 namespace gtsam {
